# apps/server/tests/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "AdminHelper Server",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/frp/generate/bulk-zip": {
      "get": {
        "description": "Generates a ZIP with frps.toml, visitor.toml and one frpc.toml per server.",
        "operationId": "gen_bulk_zip_api_frp_generate_bulk_zip_get",
        "parameters": [
          {
            "in": "query",
            "name": "config_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Gen Bulk Zip",
        "tags": [
          "frp"
        ]
      }
    },
    "/api/frp/generate/frpc-toml/{server_id}": {
      "get": {
        "operationId": "gen_frpc_toml_api_frp_generate_frpc_toml__server_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Gen Frpc Toml",
        "tags": [
          "frp"
        ]
      }
    },
    "/api/frp/generate/frps-toml": {
      "get": {
        "operationId": "gen_frps_toml_api_frp_generate_frps_toml_get",
        "parameters": [
          {
            "in": "query",
            "name": "config_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Gen Frps Toml",
        "tags": [
          "frp"
        ]
      }
    },
    "/api/frp/generate/visitor-bundle": {
      "get": {
        "description": "Returns the visitor TOML as JSON for the desktop app.\n\nThe PKI material is no longer server-minted (F2/F3: the server holds no\nsigning capability, D6). The desktop supplies its own enrolled access identity\nfor the visitor's mTLS, so the bundle carries only the TOML; ``pki`` stays\nempty for backward compatibility with the desktop's response shape.",
        "operationId": "gen_visitor_bundle_api_frp_generate_visitor_bundle_get",
        "parameters": [
          {
            "in": "query",
            "name": "config_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Gen Visitor Bundle",
        "tags": [
          "frp"
        ]
      }
    },
    "/api/frp/generate/visitor-toml": {
      "get": {
        "operationId": "gen_visitor_toml_api_frp_generate_visitor_toml_get",
        "parameters": [
          {
            "in": "query",
            "name": "config_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Config Id"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2147483647,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Gen Visitor Toml",
        "tags": [
          "frp"
        ]
      }
    },
...
  }
}